Restructure project for better maintainability
