--- app/layout.tsx
@@ -1,9 +1,0 @@
-import React from 'react'
-
-export default function RootLayout({ children }: { children: React.ReactNode }) {
-  return (
-    <html lang="en">
-      <body>{children}</body>
-    </html>
-  )
-}
