--- app/page.tsx
@@ -1,5 +1,0 @@
-import Builder from '../components/builder'
-
-export default function Home() {
-  return <Builder />
-}
