--- features/config/index.ts
@@ -1,2 +1,0 @@
-// Import/Export logic placeholder
-export {}
