--- types/index.ts
@@ -1,2 +1,0 @@
-// Shared types placeholder
-export {}
