--- utils/index.ts
@@ -1,2 +1,0 @@
-// Utility functions placeholder
-export {}
